Added Worker to Contractors
