+++ contractors/getContractors.php
@@ -1,5 +1,6 @@
 <?php
-  include "../config.php"; 
+  include "../config.php";
+  include "../utils/helper.php"; 
 
 $query = "SELECT * FROM `contractors`"; 
 
@@ -8,8 +9,8 @@ if ($result = $db->query($query)) {
 
     while($row = $result->fetch_array(MYSQLI_ASSOC)) {
             $myArray[] = $row;
-    }
-    echo json_encode($myArray);
+    }    
+    echo json_response($myArray, 200, "", null);
 }
 
 $result->close();
